Widget customizations
*in this section, I customized the widget class, customizing it's position, 
the margins, font, font color, font-size, the hover, focus, background color, text decorations, and font style. 
I did this by using the inspector tool to find the class and aligning that information with what I could find in
the parent theme to be able to accurately customize and change specific aspects of the widget class.